DEMOS-623 - stable auth, automatic user upsert, a single canonical source of user truth via UserContext, working logout, Fix profile header (#163)
* DEMOS-623: baby steps. Getting queries added

* DEMOS-623: Adding the upsert for users into local serv

Also laying foundation for a ME object to drag around

* DEMOS-623: commented  out the part that's not neeeded

* DEMOS-623: got all the user upserts working.

Now time to arrange in proper places

* DEMOS-623: add user context to create user if auth

This is replacing the static user making of profile

* DEMOS-623: removed the useUpsertUser

* DEMOS-623: fixed a config file with old cog conf

WE were using the wrang client in the backend!

* DEMOS-623: cleaned up debug code

We are getting logins and user objects

* DEMOS-623: replaced the various fetch user queries

With our new GET_CURRENT_USER context getCurrentUser

* DEMOS-623: added display and fullnme to user fetch

Also preparing this for PR. Cleaning up camel case and making sure all
the things work

* DEMOS-623: changed to getCurrentUser

instead of useCurrentUser

* DEMOS-623: fixed the header tests

* DEMOS-623: added a auth links component

Needed links, but i could not deconstruct the button base

* DEMOS-623: replaced buttons for links in header

I tried to style these guys, instead just made some link buttons.
They have hover greying. But that's it as far as styles

* DEMOS-623: correcting and updating tests

Lotta stuff changed, so let's get these tests working

* DEMOS-623: corrected and update unit test.

1. Now we do not need uerid in many place.
2. We have login in and we can mock that.
3. Needed to make a duper request in mocks for headers to work

* DEMOS-623: removed a query we weren't using

* DEMOS-623: added a test for sign in links

* DEMOS-623: removed debug code

* DEMOS-623: removed dep that is unused

* DEMOS-623: removed comments

* DEMOS-6213: removed debug code

* DEMOS-623: update env example and clean up

* DEMOS-623: removed deubg code form auth utils

* DEMOS-623: cleaned up auth significantly.

1. third locationed the user create
2. Added a HTTP context for local
3. Added GraphQLContext to lamda entry point.
NOTE Refactoed the heck out of this to harden the auth

* DEMOS-623: nuked some comments, lil clean up

* DEMOS-623: nuking comments

* DEMOS-623: comment nuking binge

* DEMOS-623: nuked the children in auth buttons

* DEMOS-623: replaced a mock provider with our Prov

* DEMOS-623: added lambda check for decoded headers

We check the claim from authorizor for plain text headers.
Specifically event.requestContext.authorizer
Then, if not found, we decode

* DEMOS-623: cleaned up some of the formatting

* DEMOS-623: removed userid from a head comp

This userid was not in use. We eliminated the use for it

* DEMOS-623: added data-testid to Create new Button

Also minor clean up, turned dupe user mock into a todo

* DEMOS-623: fixed issue with unused attrib _info

in resolver

* DEMOS-623: fixing linter error

* DEMOS-623: fix of package-lock file

* DEMOS-623: created new lock file after merge