Demos 595: removed faker_data and moved type data to constants (#240)
* DEMOS-595: added demo statuses, nukes fake data

1. Worked it out to where i am replacing the demo statuses

* DEMOS-595: added the server/constants to demo stts

1. Using constants.ts to replace the frontend classes

* DEMOS-595: added states to constants and more

* DEMOS-595: Testing demo was pending and not active

Per the mock data i had to make this change.

* DEMOS-595: made some adjustments to status

Since they changes a bit, there is no pending, then we had to make some
changes to tests etc.

* DEMOS-595: cleaned up some typo'd statues

There are a lot of statuses everywhere.

* DEMOS-595: Addressing centralization of demo stus

* DEMOS-595: added seeds from constants

* DEMOS-595: added a type to types list

Also instead of mock the states we just pull the states from the
constants. PR find.

* DEMOS-595: removed the mockStatuses, not needed