Demos 529 Validate User has a role, also added the name hydration in database from cognito (#250)
* DEMOS-529: checking main

* DEMOS-529: fixed the auth link signout button.

It was not recognizing clicks in the box.

* DEMOS-529: added method to verify role

This is a bit hard due to the logic. But will need a way to reflect a
login

* DEMOS-529: added redirect for bad role.

Since this should be very rare we just redirec the user back to login
screen

* DEMOS-529: added hydration of person's names

Not exactly in scope. but easy to knock out.

* DEMOS-529: removed TODO and cleaned up displayname

Since these fields cannot be null, then let's reduce the "backup" names

* DEMOS-529: remove todo, and fixed options

* DEMOS-529: fixed up breaking tests

Mostly around update name in the header.

* DEMOS-529: added refactor to auth.utils

Also moved the 404 into the NOT development per suggestion

* DEMOS-529: removed the returns in a compoenent

* DEMOS-529: added the config update for test

* DEMOS-529: adjusted cognito config for test

* DEMOS-529: added some console logs/errors

We do not have a lot of logs in the code, so lets add some while i
remember what everything does here

* DEMOS-529: removed the non-login person type

* DEMOS-529: added IDM post_logout uri to deploypipe

* DEMOS-529: added idm logout url correctly to deply

* DEMOS-529: cleaned up the logic flow a bit

* DEMOS-529: one more cleanup with localhost