Demos 756 b - Completeness phase integration.  (#417)
* DEMOS-756: adjustments to complete completeness

We need to ship 3 dates with application IDs to the server via mutator.

* DEMOS-756: got this as finished as i can.

Docuemnt upload blocking

* DEMOS-756: down to a date error

* DEMOS-756: replaced useEffect with better useStates

* DEMOS-756: added Completeness Docuemnt Seed

Since the doc upload is borked, imma just seed some

* DEMOS-756: got the incomplete, complete set up

It's jsut a query. That makes way a lot of sense

* DEMOS-756: mostly done. Working on unit test.

* DEMOS-756: merged main couple update

* DEMOS-756: updated unit tests, incomplete tests

I had AI generate the tests, but i had to go back thru and adjust
everything.

* DEMOS-756: resettting the commented out validation

Thought this really is not very forgiving. so imma havfa mess with this
later.

* DEMOS-756: some minor updates CONST start/end date

This is for better adjustability and conistency

* DEMOS-756: aded description to doc stated "seeded"

This way perhaps we won't forget to fix it. Ticket generated.

* DEMOS-756: couple linter fixes