Demos 1055 Local docs: Write directly to docs table when env set. Allows for local dev without localstack, allows for localstack and leave uppers untouched.   (#456)
* DEMOS-1055: docs take 2: added bypass for S3

Just seems like a lot of moving parts to keep involving the S3
localstack stuff. So let's just make upload simple for local

* DEMOS-1055: state where docs are working locally

* DEMOS-1055: local doc upload working

There are some refetches that are not however.

* DEMOS-1055: added refetch to phase 3.

This was added doc will show up at least in one place

* DEMOS-1055: simplified the resolver

* DEMOS-1055: stopped working, See last commit

trying leverage the url instead of the bypass node

* DEMOS-1055: doc upload works locally now!

We short cut the document upload process. On local. It just writes to
document table

* DEMOS-1055: added some phase_id to completeness.

the same thing can be copied over to other phases but sets the phase_id
in docuemnts

* DEMOs-1055: imported query from app workflow