Demos 1046 remove hooks (#520)
* DEMOS-1046: first pass at hook removal.

Knocked out useDemo, useDialogForm, useDateVal. Able to just delete
most, one function i had to move into it's scope as was not re-used

* DEMOS-1046: removed all the unused hooks.

There was one in use. GET_DEMONSTRATION_BY_ID_QUERY and i moved that to
DemoDetails page. that seems to be the best candidate for a demo query,

* DEMOS-1046: move use demo tests to demoDeets

Since the query was moved, let's jsut move the tests over too.
This may be overlap in some case, but it's not blatant.