DEMOS-975: Minimal centralized date picker, and seeder stuff (#528)
* DEMOS-975: made a centralized date picker.

* DEMOS-975: Removed all the comments

* DEMOS-975: removing the ... spread operator

So that we can't just add like the kitchen sink to this, thus most
configuring has to be done in the component and not where it's used.

* DEMOS-975: Reduced the footprint of DatePicker

Also cleaned up implementation in column fitler

* DEMOS-975: merged my doc seeder with exists.

The issue i am trying to fix is that the document dates are all the
same, So hard to test filter.

* DEMOS-975: removed complexity, and remove ...props

Could be candidate for "and the rest of the input stuff" but not yet

* DEMOS-975: removed comment

* DEMOS-975: simplest possibel verison

* DEMOS-975: reduction in comp to make simple

* DEMOS-975: adjusted so we do not have dupe docs

but i did add random dates to test search functionality in doc table

* DEMOS-975: missed a doc addition